refactor from JPA Query Method to JPA Specification

+++ src/main/java/com/langthang/model/dto/response/AccountDTO.java
@@ -22,7 +22,7 @@ public class AccountDTO {
 
     private String slug;
 
-    private int postCount;
+    private long postCount;
 
     private int followCount;
